Staff list 80% complete
The table of staff has been expanded on the backend. and the View can now show the nessessary information. no UI has been implemented on the buttons that were added [View Activity] and [Edit]. The Joined date value is in isostring format so that hasnt been fixed yet. The database still is composed of Role and Staff