Land registration, View Land,
The image view for land is fixed.The Land registration is done. The View for the Land is partially done ... The only thing that remains is the dropdown. The Staff update is started but not yet done. The Staff update status is not implemeted yet.. and no archive is implemented